+++ CSharpAPI/Controllers/ItemsController.cs
@@ -1,6 +1,8 @@
 using Microsoft.AspNetCore.Mvc;
 using CSharpAPI.Models;
 using CSharpAPI.Service;
+using CSharpAPI.Models.Auth;
+using CSharpAPI.Services.Auth;
 
 namespace CSharpAPI.Controllers
 {
@@ -9,15 +11,26 @@ namespace CSharpAPI.Controllers
     public class ItemsController : ControllerBase
     {
         private readonly IItemsService _itemsService;
+        private readonly IAuthService _authService;
 
-        public ItemsController(IItemsService itemsService)
+        public ItemsController(IItemsService itemsService, IAuthService authService)
         {
             _itemsService = itemsService;
+            _authService = authService;
+        }
+
+        private async Task<bool> CheckAccess(string method)
+        {
+            var user = HttpContext.Items["User"] as ApiUser;
+            return await _authService.HasAccess(user, "items", method);
         }
 
         [HttpGet("all")]
         public async Task<IActionResult> GetAllItems()
         {
+            if (!await CheckAccess("GET"))
+                return Forbid();
+
             var _items = await _itemsService.GetAllItems();
             return Ok(_items);
         }
@@ -25,6 +38,9 @@ namespace CSharpAPI.Controllers
         [HttpGet("{uid}")]
         public async Task<IActionResult> GetItemById(string uid)
         {
+            if (!await CheckAccess("GET"))
+                return Forbid();
+
             var _item = await _itemsService.GetItemById(uid);
             if (_item == null) return NotFound($"Item with uid {uid} not found.");
             return Ok(_item);
@@ -33,6 +49,9 @@ namespace CSharpAPI.Controllers
         [HttpPost]
         public async Task<IActionResult> CreateItem(ItemModel item)
         {
+            if (!await CheckAccess("POST"))
+                return Forbid();
+
             if (item == null) return BadRequest("Request is empty!");
             await _itemsService.CreateItem(item);
             return CreatedAtAction(nameof(GetItemById), new { uid = item.uid }, item);
@@ -41,6 +60,9 @@ namespace CSharpAPI.Controllers
         [HttpPut("{uid}")]
         public async Task<IActionResult> UpdateItem(string uid, ItemModel item)
         {
+            if (!await CheckAccess("PUT"))
+                return Forbid();
+
             if (item == null) return BadRequest("Request is empty!");
             await _itemsService.UpdateItem(uid, item);
             return Ok($"Item {uid} has been updated!");
@@ -49,6 +71,9 @@ namespace CSharpAPI.Controllers
         [HttpDelete("{uid}")]
         public async Task<IActionResult> DeleteItem(string uid)
         {
+            if (!await CheckAccess("DELETE"))
+                return Forbid();
+
             await _itemsService.DeleteItem(uid);
             return Ok("Item has been deleted!");
         }
@@ -56,6 +81,9 @@ namespace CSharpAPI.Controllers
         [HttpGet("line/{lineId}")]
         public async Task<IActionResult> GetItemsByLineId(int lineId)
         {
+            if (!await CheckAccess("GET"))
+                return Forbid();
+
             var _itemline = await _itemsService.GetItemsByLineId(lineId);
             if (_itemline == null) return NotFound("Itemline not found!");
             return Ok(_itemline);
@@ -64,6 +92,9 @@ namespace CSharpAPI.Controllers
         [HttpGet("group/{groupId}")]
         public async Task<IActionResult> GetItemsByGroupId(int groupId)
         {
+            if (!await CheckAccess("GET"))
+                return Forbid();
+
             var _itemgroup = await _itemsService.GetItemsByGroupId(groupId);
             if (_itemgroup == null) return NotFound("Itemgroup not found!");
             return Ok(_itemgroup);
@@ -72,10 +103,12 @@ namespace CSharpAPI.Controllers
         [HttpGet("supplier/{supplierId}")]
         public async Task<IActionResult> GetItemsBySupplierId(int supplierId)
         {
+            if (!await CheckAccess("GET"))
+                return Forbid();
+
             var _supplierId = await _itemsService.GetItemsBySupplierId(supplierId);
             if (_supplierId == null) return NotFound("Supplier not found!");
             return Ok(_supplierId);
         }
     }
-
 }
